requirements.txt
langchain
langchain-core
langchain-community
langchain-nvidia-ai-endpoints
transformers
langsmith
streamlit
beautifulsoup4
lxml
python-dotenv
